+++ pkg/deployments/CHANGELOG.md
@@ -13,6 +13,8 @@
 ### API Changes
 
 - Made `getBalancerContractAbi`, `getBalancerContractBytecode`, `getBalancerContractAddress` and `getBalancerDeployment` synchronous rather than asynchronous functions.
+- Added `getBalancerContractArtifact` which returns a artifact file for the contract in the same format used by Hardhat.
+- Deprecated `getBalancerContractBytecode` in favour of `getBalancerContractArtifact`.
 
 ## 3.0.0 (2022-10-25)